feat: Move test-related functions from test_runner.py to tests/utils.py
