+++ frappe/tests/test_global_search.py
@@ -4,8 +4,8 @@
 import frappe
 from frappe.custom.doctype.property_setter.property_setter import make_property_setter
 from frappe.desk.page.setup_wizard.install_fixtures import update_global_search_doctypes
-from frappe.test_runner import make_test_objects
 from frappe.tests import IntegrationTestCase
+from frappe.tests.utils import make_test_objects
 from frappe.utils import global_search, now_datetime
