+++ frappe/tests/test_goal.py
@@ -2,8 +2,8 @@
 # License: MIT. See LICENSE
 
 import frappe
-from frappe.test_runner import make_test_objects
 from frappe.tests import IntegrationTestCase
+from frappe.tests.utils import make_test_objects
 from frappe.utils import format_date, today
 from frappe.utils.goal import get_monthly_goal_graph_data, get_monthly_results
