+++ frappe/tests/test_permissions.py
@@ -22,8 +22,8 @@ from frappe.permissions import (
 	remove_user_permission,
 	update_permission_property,
 )
-from frappe.test_runner import make_test_records_for_doctype
 from frappe.tests import IntegrationTestCase
+from frappe.tests.utils import make_test_records_for_doctype
 from frappe.utils.data import now_datetime
 
 test_dependencies = ["Blogger", "Blog Post", "User", "Contact", "Salutation"]
