+++ frappe/tests/test_sitemap.py
@@ -5,7 +5,7 @@ from frappe.utils import get_html_for_route
 
 class TestSitemap(IntegrationTestCase):
 	def test_sitemap(self):
-		from frappe.test_runner import make_test_records
+		from frappe.tests.utils import make_test_records
 
 		make_test_records("Blog Post")
 		blogs = frappe.get_all("Blog Post", {"published": 1}, ["route"], limit=1)
